+++ packages/core/decorators/exception-filters.decorator.ts
@@ -1,5 +1,5 @@
 import { ExceptionFilter } from "@venok/core/interfaces/features/exception-filter.interface";
-import { isFunction } from "@venok/core/utils/shared.utils";
+import { isFunction } from "@venok/core/helpers/shared.helper";
 import { EXCEPTION_FILTERS_METADATA } from "@venok/core/constants";
 import { validateEach } from "@venok/core/helpers/validate-each.helper";
 import { extendArrayMetadata } from "@venok/core/helpers/extends-metadata.helper";
