+++ packages/core/decorators/use-guards.decorator.ts
@@ -1,5 +1,5 @@
 import { CanActivate } from "@venok/core/interfaces/features/guards.interface";
-import { isFunction } from "@venok/core/utils/shared.utils";
+import { isFunction } from "@venok/core/helpers/shared.helper";
 import { GUARDS_METADATA } from "@venok/core/constants";
 import { validateEach } from "@venok/core/helpers/validate-each.helper";
 import { extendArrayMetadata } from "@venok/core/helpers/extends-metadata.helper";
