+++ packages/core/decorators/use-pipes.decorator.ts
@@ -1,5 +1,5 @@
 import { PipeTransform } from "@venok/core/interfaces/features/pipes.interface";
-import { isFunction } from "@venok/core/utils/shared.utils";
+import { isFunction } from "@venok/core/helpers/shared.helper";
 import { PIPES_METADATA } from "@venok/core/constants";
 import { extendArrayMetadata } from "@venok/core/helpers/extends-metadata.helper";
 import { validateEach } from "@venok/core/helpers/validate-each.helper";
