+++ packages/core/discovery/module.ts
@@ -1,6 +1,6 @@
 import { MetadataScanner } from "@venok/core/metadata-scanner";
 import { Module } from "@venok/core/decorators/module.decorator";
-import { DiscoveryService } from "@venok/core/discovery/discovery-service";
+import { DiscoveryService } from "@venok/core/discovery/service";
 
 /**
  * @publicApi
