+++ packages/core/errors/messages.ts
@@ -1,6 +1,6 @@
 import { InjectorDependency, InjectorDependencyContext } from "@venok/core/injector/injector";
 import { DynamicModule, ForwardReference } from "@venok/core/interfaces/modules";
-import { isNull, isSymbol } from "@venok/core/utils/shared.utils";
+import { isNull, isSymbol } from "@venok/core/helpers/shared.helper";
 import { Module } from "@venok/core/injector/module/module";
 import { Type } from "@venok/core/interfaces";
