+++ packages/core/helpers/random-string-generator.helper.ts
