+++ packages/core/hooks/before-app-shutdown.hook.ts
@@ -1,4 +1,4 @@
-import { isFunction, isNull } from "@venok/core/utils/shared.utils";
+import { isFunction, isNull } from "@venok/core/helpers/shared.helper";
 import { InstanceWrapper } from "@venok/core/injector/instance/wrapper";
 import { Module } from "@venok/core/injector/module/module";
 import { InjectionToken } from "@venok/core/interfaces";
