+++ packages/core/hooks/on-app-bootstrap.hook.ts
@@ -1,4 +1,4 @@
-import { isFunction, isNull } from "@venok/core/utils/shared.utils";
+import { isFunction, isNull } from "@venok/core/helpers/shared.helper";
 import { InstanceWrapper } from "@venok/core/injector/instance/wrapper";
 import { Module } from "@venok/core/injector/module/module";
 import { getNonTransientInstances, getTransientInstances } from "@venok/core/helpers/transient.helper";
