+++ packages/core/hooks/on-module-destroy.hook.ts
@@ -2,7 +2,7 @@ import { InjectionToken } from "@venok/core/interfaces";
 import { InstanceWrapper } from "@venok/core/injector/instance/wrapper";
 import { getNonTransientInstances, getTransientInstances } from "@venok/core/helpers/transient.helper";
 import { Module } from "@venok/core/injector/module/module";
-import { isFunction, isNull } from "@venok/core/utils/shared.utils";
+import { isFunction, isNull } from "@venok/core/helpers/shared.helper";
 import { OnModuleDestroy } from "@venok/core/interfaces/hooks";
 
 /**
