+++ packages/core/injector/container.ts
@@ -13,7 +13,7 @@ import { EnhancerSubtype, GLOBAL_MODULE_METADATA, REQUEST } from "@venok/core/co
 import { ContextId } from "@venok/core/injector/instance/wrapper";
 import { ApplicationConfig } from "@venok/core/application/config";
 import { InitializeOnPreviewAllowlist } from "@venok/core/inspector/initialize-on-preview.allowlist";
-import { DiscoverableMetaHostCollection } from "@venok/core/discovery/discoverable-meta-host-collection";
+import { DiscoverableMetaHostCollection } from "@venok/core/discovery/meta-host-collection";
 import { SerializedGraph } from "@venok/core/inspector/serialized-graph";
 import { InternalCoreModule } from "@venok/core/injector/internal-core-module/internal-core-module";
