+++ packages/core/injector/instance/wrapper.ts
@@ -4,9 +4,9 @@ import { EnhancerSubtype } from "@venok/core/constants";
 import { Scope, Type } from "@venok/core/interfaces";
 import { SettlementSignal } from "@venok/core/injector/settlement-signal";
 import { STATIC_CONTEXT } from "@venok/core/injector/constants";
-import { isNull, isString, isUndefined } from "@venok/core/utils/shared.utils";
+import { isNull, isString, isUndefined } from "@venok/core/helpers/shared.helper";
 import { UuidFactory } from "@venok/core/helpers/uuid.helper";
-import { randomStringGenerator } from "@venok/core/utils/random-string-generator.util";
+import { randomStringGenerator } from "@venok/core/helpers/random-string-generator.helper";
 import { colors } from "@venok/core/helpers/color.helper";
 import { isClassProvider, isFactoryProvider, isValueProvider } from "@venok/core/injector/helpers/classifier.helper";
 import { Logger, LoggerService } from "@venok/core/services/logger.service";
