+++ packages/core/pipes/consumer.ts
