+++ packages/core/test/pipes/consumer.spec.ts
